Uploaded "A03 Magnetometer Analysis"
Using sample data from public flights, an offline Kalman Filter is tested to remove hard-iron and soft-iron offsets from the magnetometer. It is the stepstone of a final version that will be used to test the best calibration method.

This program must be tested using different positions of the plane to understand what is the most efficient calibration procedure. Possible alternatives include:
- Real-time calibration (doesn't seem the most effective, proven in this data study) 
- 360 degrees rotations before the start of the flight (potential solution)
- 3D 360 degrees rotations (unpractical for medium size planes)