Bluetooth Control
@Shiro-01 I attach you as a reviewer of this commit for you to test the program with the full prototype. However, the changes are done mostly in a safe way with the use of the macro "WITH_BLUETOOTH", and I am merging the branch to save your reviewing time and avoid outdated pull requests.

The code implements BLE functionalities without the use of the mentioned RemoteXY library, which would interfere with some natural RTOS functions. A downside may be limited GUI, but this achieves simple and flexible functionality that can easily be extended to future needs.

How to connect with the program?
1) Download "nRF Connect" compatible with iOS and Android
2) Connect with device "NFC25-Datalogger"
3) Inside "Advertised Services", write "01" in the only available one
Every "01" sent is an hexadecimal value that is programmed to be understood as a button.